rename some files, refactor toward more non-event sourced command scheduling tests for the SQL scheduler
