+++ Domain.Sql.Tests/Domain.Sql.Tests.csproj
@@ -150,7 +150,9 @@
     <Compile Include="SqlCommandSchedulerIdempotencyTests_EventSourced.cs" />
     <Compile Include="SqlCommandSchedulerIdempotencyTests_NonEventSourced.cs" />
     <Compile Include="SqlCommandSchedulerTests.cs" />
+    <Compile Include="SqlCommandSchedulerTests_EventSourced.cs" />
     <Compile Include="SqlCommandSchedulerTests_Legacy.cs" />
+    <Compile Include="SqlCommandSchedulerTests_NonEventSourced.cs" />
     <Compile Include="SqlCommandSchedulerTests_Pipeline.cs" />
     <Compile Include="SqlEventSourcedRepositoryMigrationTests.cs" />
     <Compile Include="SqlEventSourcedRepositoryTests.cs" />
