+++ Domain.Sql.Tests/SqlCommandSchedulerTests_NonEventSourced.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft. All rights reserved. 
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Its.Domain.Sql.Tests
+{
+    public abstract class SqlCommandSchedulerTests_NonEventSourced : SqlCommandSchedulerTests
+    {
+    }
+}
