+++ Domain.Sql.Tests/SqlCommandSchedulerTests_Pipeline.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft. All rights reserved. 
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
 using System.Linq;
 using Microsoft.Its.Domain.Sql.CommandScheduler;
@@ -7,7 +10,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
 {
     [Category("Command scheduling")]
     [TestFixture]
-    public class SqlCommandSchedulerTests_Pipeline : SqlCommandSchedulerTests
+    public class SqlCommandSchedulerTests_Pipeline : SqlCommandSchedulerTests_EventSourced
     {
         protected override void ConfigureScheduler(Configuration configuration)
         {
